refactor(settings): move terminal link actions row to Terminal settings
The "Show terminal link actions" row toggles terminalLinkActionPopoverEnabled,
which has only ever governed terminal links, but it rendered in the Browser
pane nested under Link Routing. Render it in Terminal -> Terminal Interaction
instead.

The copy moves out of browser-link-routing-copy.ts into a terminal-owned
terminal-link-actions-copy.ts, the deep-link target id becomes
terminal-link-actions, and the row leaves the Browser search index, which
shifts the positional indices BrowserPane.tsx uses to gate later rows.
Translation keys stay as they are so no catalog entry moves.
