+++ src/renderer/src/components/settings/TerminalInteractionSection.tsx
@@ -7,6 +7,7 @@ import { SettingsSubsectionHeader, SettingsSwitchRow } from './SettingsFormContr
 import { SearchableSetting } from './SearchableSetting'
 import { matchesSettingsSearch } from './settings-search'
 import { getTerminalRightClickToPasteSearchEntry } from './terminal-windows-search'
+import { TerminalLinkActionsSetting } from './TerminalLinkActionsSetting'
 import { OSC52_CLIPBOARD_SETTING_ID } from '../terminal-pane/osc52-clipboard-setting-anchor'
 import { isMacPlatform } from '../terminal-pane/terminal-link-open-hints'
 import { translate } from '@/i18n/i18n'
@@ -268,6 +269,12 @@ export function TerminalInteractionSection({
             />
           </SearchableSetting>
         ) : null}
+
+        <TerminalLinkActionsSetting
+          settings={settings}
+          isMac={isMac}
+          updateSettings={updateSettings}
+        />
 
         <SearchableSetting
           title={translate(
