+++ src/renderer/src/components/settings/browser-search.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from 'vitest'
 
 import ko from '@/i18n/locales/ko.json'
 import { i18n } from '@/i18n/i18n'
-import { getBrowserPaneSearchEntries, getTerminalLinkActionSearchKeywords } from './browser-search'
+import { getBrowserPaneSearchEntries } from './browser-search'
 import {
   getBrowserLinkRoutingDescription,
   getBrowserLinkRoutingShortcutLabel,
@@ -47,17 +47,6 @@ describe('browser settings search copy', () => {
     expect(linkRoutingEntry?.description).toBe(getBrowserLinkRoutingDescription({ isMac: false }))
     expect(linkRoutingEntry?.keywords).toContain('ctrl')
     expect(linkRoutingEntry?.keywords).not.toContain('cmd')
-
-    const terminalActionsEntry = getBrowserPaneSearchEntries({ isMac: false }).find(
-      (entry) => entry.title === 'Show terminal link actions'
-    )
-    expect(terminalActionsEntry?.description).toContain('Ctrl-click')
-    expect(terminalActionsEntry?.description).not.toContain('Cmd/Ctrl')
-    expect(terminalActionsEntry?.keywords).toEqual(
-      getTerminalLinkActionSearchKeywords({ isMac: false })
-    )
-    expect(terminalActionsEntry?.keywords).toContain('browser')
-    expect(terminalActionsEntry?.keywords).toContain('ctrl')
   })
 
   // Why: shipping the opt-in must not reword this row for anyone who never enables
@@ -90,7 +79,6 @@ describe('browser link routing modifier copy', () => {
       'Default Zoom',
       'Link Routing',
       'Hold Shift to open in Orca',
-      'Show terminal link actions',
       'Localhost Worktree Labels',
       'Session & Cookies'
     ])
