+++ src/renderer/src/lib/settings-navigation-types.ts
@@ -61,7 +61,7 @@ const SETTINGS_NAV_INTENT_SET: ReadonlySet<string> = new Set(SETTINGS_NAV_INTENT
 
 export type SettingsNavTarget = (typeof SETTINGS_NAV_TARGETS)[number]
 export const FULL_DISK_ACCESS_SETTINGS_TARGET_ID = 'developer-permissions-full-disk-access'
-export const BROWSER_TERMINAL_LINK_ACTIONS_SETTINGS_TARGET_ID = 'browser-terminal-link-actions'
+export const TERMINAL_LINK_ACTIONS_SETTINGS_TARGET_ID = 'terminal-link-actions'
 export const GLOBAL_WORKTREE_VISIBILITY_SETTINGS_TARGET_ID = 'general-global-worktree-visibility'
 
 export type SettingsNavigationTarget = {
